+++ src/tags_int.hpp
@@ -344,7 +344,7 @@ namespace Exiv2 {
             return os;
         }
 
-        for (int i=0; i< value.count(); i++) {
+        for (size_t i=0; i< value.count(); i++) {
             if (i != 0)
                 os << ", ";
             const TagVocabulary* td = find(array, value.toString(i));
